+++ src/seller/seller.service.ts
@@ -23,11 +23,25 @@ export class SellerService {
   }
 
   async deleteSeller(sellerId: number) {
+    const seller = await prisma.seller.findUnique({
+      where: { id: sellerId },
+    });
+    if (seller == null) {
+      throw new NotFoundException('Seller not found');
+    }
     await prisma.seller.delete({ where: { id: sellerId } });
   }
 
   async getSeller(sellerId: number) {
-    return await prisma.seller.findUnique({ where: { id: sellerId } });
+    const seller = await prisma.seller.findUnique({
+      where: { id: sellerId },
+    });
+    if (seller == null) {
+      throw new NotFoundException('Seller not found');
+    }
+    return {
+      seller: seller,
+    };
   }
 
   async updateSeller(sellerId: number, updateSellerDto: UpdateSellerDto) {
